size-opt: Compress static vector icon coordinates to int16_t to save 54KB binary size

+++ QuickView/GeekIconLibrary.h
@@ -14,7 +14,7 @@ namespace GeekIcons {
     // --- Vector Infrastructure ---
     struct IconPathCommand {
         char type; // 'M', 'L', 'B', 'Z'
-        float x1, y1, x2, y2, x3, y3;
+        int16_t x1, y1, x2, y2, x3, y3;
     };
 
     struct VectorIcon {
